windows compilation fixes

+++ src/screen.cpp
@@ -519,7 +519,7 @@ void Screen::initialize(GLFWwindow *window, bool shutdown_glfw) {
     __nanogui_screens[m_glfw_window] = this;
 
     for (size_t i = 0; i < (size_t) Cursor::CursorCount; ++i)
-        m_cursors[i] = glfwCreateStandardCursor(GLFW_ARROW_CURSOR + i);
+        m_cursors[i] = glfwCreateStandardCursor(GLFW_ARROW_CURSOR + (int) i);
 
     /// Fixes retina display-related font rendering issue (#185)
     nvgBeginFrame(m_nvg_context, m_size[0], m_size[1], m_pixel_ratio);
